TRANSFERENCEIA DE BANCO DE DADOS/AJUSTES NO CADASTRO DE USUARIO

+++ PDV/src/main/java/br/com/pdv/bean/ProdutoBean.java
@@ -113,7 +113,7 @@ public class ProdutoBean implements Serializable {
 		try {
 			String valor = FacesUtil.getParam("produtoCod");
 			if (valor != null) {
-				Integer codigo = Integer.parseInt(valor);
+				Long codigo = Long.parseLong(valor);
 
 				this.produto = produtoDAO.buscarPeloCodigo(codigo);
 			}
